+++ addon/services/portal-service.js
@@ -293,6 +293,14 @@ export default Service.extend(serviceMixin, {
   getSigninSettings (portalOpts) {
     const urlPath = `/portals/self/signinSettings?f=json`;
     return this.request(urlPath, {}, portalOpts);
+  },
+
+  updateContacts (contacts, portalOpts) {
+    /*
+      contacts: [contact1Username, contact2Username]
+    */
+    const urlPath = `/portals/self/update?contacts=${contacts}&f=json`;
+    return this._post(urlPath, null, portalOpts);
   }
 
 });