+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 2.6.0
 ### Added
 - `removeSocialProviders` call to `portal-service` to remove social providers
+- `updateContacts` call to `portal-service` to update support contacts
 
 ## 2.5.0
 ### Added